Move concluding to concluding/instigator
