+++ packages/wallet/src/redux/protocols/concluding/instigator/components/wait-for-opponent-conclude.tsx
